resolve feedback for BlobByteArray code review
- remove toString() override from BlobByteArray to prevent excessive logging
- extract PostgreSqlJdbcBlobByteArrayType to separate file
- remove duplicate type registration in MySqlJdbcDataTypeProvider
