+++ komapper-core/src/main/kotlin/org/komapper/core/type/BlobByteArray.kt
@@ -1,8 +1,4 @@
 package org.komapper.core.type
 
 @JvmInline
-value class BlobByteArray(val value: ByteArray) {
-    override fun toString(): String {
-        return value.contentToString()
-    }
-}
+value class BlobByteArray(val value: ByteArray)
