+++ komapper-dialect-mysql-jdbc/src/main/kotlin/org/komapper/dialect/mysql/jdbc/MySqlJdbcDataTypeProvider.kt
@@ -43,9 +43,8 @@ class MySqlJdbcDataTypeProvider(next: JdbcDataTypeProvider) : AbstractJdbcDataTy
             JdbcBooleanType("bit(1)"),
             JdbcByteType("tinyint"),
             JdbcByteArrayType("varbinary(500)"),
-            JdbcDoubleType("double precision") { toDoubleLiteral(it) },
             JdbcBlobByteArrayType("blob"),
-            JdbcDoubleType("double precision"),
+            JdbcDoubleType("double precision") { toDoubleLiteral(it) },
             JdbcClobType("text"),
             JdbcClobStringType("text"),
             JdbcFloatType("real"),
