+++ komapper-dialect-postgresql-jdbc/src/main/kotlin/org/komapper/dialect/postgresql/jdbc/PostgreSqlJdbcDataTypeProvider.kt
@@ -1,9 +1,7 @@
 package org.komapper.dialect.postgresql.jdbc
 
-import org.komapper.core.type.BlobByteArray
 import org.komapper.dialect.postgresql.PostgreSqlLiteral.toDoubleLiteral
 import org.komapper.dialect.postgresql.PostgreSqlLiteral.toOffsetDateTimeLiteral
-import org.komapper.jdbc.AbstractJdbcDataType
 import org.komapper.jdbc.AbstractJdbcDataTypeProvider
 import org.komapper.jdbc.JdbcArrayType
 import org.komapper.jdbc.JdbcBigDecimalType
@@ -30,10 +28,6 @@ import org.komapper.jdbc.JdbcStringType
 import org.komapper.jdbc.JdbcUByteType
 import org.komapper.jdbc.JdbcUIntType
 import org.komapper.jdbc.JdbcUShortType
-import java.sql.JDBCType
-import java.sql.PreparedStatement
-import java.sql.ResultSet
-import kotlin.reflect.typeOf
 import kotlin.time.ExperimentalTime
 
 class PostgreSqlJdbcDataTypeProvider(next: JdbcDataTypeProvider) :
@@ -47,7 +41,7 @@ class PostgreSqlJdbcDataTypeProvider(next: JdbcDataTypeProvider) :
             JdbcBooleanType("boolean"),
             JdbcByteType("smallint"),
             JdbcByteArrayType("bytea"),
-            PostgreSqlJdbcBlobByteArrayType,
+            PostgreSqlJdbcBlobByteArrayType(),
             JdbcClobStringType("text"),
             JdbcDoubleType("double precision") { toDoubleLiteral(it) },
             JdbcFloatType("real"),
@@ -67,19 +61,5 @@ class PostgreSqlJdbcDataTypeProvider(next: JdbcDataTypeProvider) :
             JdbcUShortType("integer"),
             PostgreSqlJdbcUUIDType,
         )
-    }
-}
-
-object PostgreSqlJdbcBlobByteArrayType :
-    AbstractJdbcDataType<BlobByteArray>(typeOf<BlobByteArray>(), JDBCType.BINARY) {
-    override val name: String = "bytea"
-    override fun doGetValue(rs: ResultSet, index: Int): BlobByteArray? {
-        return rs.getBytes(index)?.let { BlobByteArray(it) }
-    }
-    override fun doGetValue(rs: ResultSet, columnLabel: String): BlobByteArray? {
-        return rs.getBytes(columnLabel)?.let { BlobByteArray(it) }
-    }
-    override fun doSetValue(ps: PreparedStatement, index: Int, value: BlobByteArray) {
-        ps.setBytes(index, value.value)
     }
 }
